+++ changelog.md
@@ -2,6 +2,7 @@
 ### Changes
 - fixed recipe to upgrade sack into lunchbox, now it properly makes items stay and refunds incompatible items.
 - fixed recipe to upgrade sack into toolbelt, now it properly makes items stay and refunds incompatible items.
+- fixed recipe to upgrade sack into backpacks, now it properly makes items stay and refunds incompatible items.
 
 ## [0.0.33]
 ### Changes